feat: ✨ update auth middlewares to allow other auth middlewares (#469)
## Description

The main purpose of this PR is to give room to custom authentication
middlewares in the projects.

The middlewares were cutting off the execution and giving forbidden
errors other than the `Bearer` and `Basic` auth token types.

Also getting the `token_type` and `token` from `HTTP_AUTHORIZATION` is
made a utility function to keep the code DRY.