fix(rollbar): Update 2.26.x version handling in Rollbar templates (#486)
There was a bug that prevented the loading of `rollbar.min.js` file on
Django app.

![Screenshot 2025-06-10 at 11 31
21 PM](https://github.com/user-attachments/assets/c5af2f2f-4245-4126-8cc4-b9bfb78126a3)

In the [bawshuman-django
repo](https://github.com/BawsHuman/bawshuman-django/blob/224068e4c36bcbee88f5ad4a56f21c4f44d8f563/baws/templates/baws/fragments/js/common.html#L1),
we specify the full version "2.26.4". But in the rollbar.html file, we
are doing a strict equality check to "2.26". This causes the `v1.x.html`
template to load instead of `v2.26.x.html`, causing the wrong CDN URL to
be used.

This fix allows specifying the version for 2.26.x with either "2.26"
which will fallback to "2.26.0", or you can specify a specific patch
such as "2.26.4".